LPD-31842 Fix tests: move the model search configurator to the integration test to limit the scope
